cs/server/requirements.txt
# Ocean Plastic Forecast - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Data Processing
numpy==1.25.2
pandas==2.1.3
scipy==1.11.4

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Container Management
docker==6.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
